% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlineardata.R
\docType{data}
\name{nonlineardata}
\alias{nonlineardata}
\title{nonlineardata}
\format{
A data.frame with 3733 observations on 9 variables:
\itemize{
 \item \strong{Y:} The incidence of cardiovascular diseases.
 \item \strong{bmi:} The BMI level.
 \item \strong{insulin:} The insulin level.
 \item \strong{Z.1:} SNP genotypes.
 \item \strong{Z.2:} SNP genotypes.
 \item \strong{Z.3:} SNP genotypes.
 \item \strong{Z.4:} SNP genotypes.
 \item \strong{age:} the age of the subject.
 \item \strong{sex:} the sex of the subject.
}
}
\source{
The Framingham Heart Study data supported by the National Heart, Lung, and Blood Institute (NHLBI) in collaboration with Boston University.
}
\usage{
data(nonlineardata)
}
\description{
Psuedo data provided by Youjin Lee, which is generated mimicing the structure of Framingham Heart Study data.
}
\examples{
data(nonlineardata)
}
\keyword{datasets}
