% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretest.R
\name{summary.pretest}
\alias{summary.pretest}
\title{Summary of pretest}
\usage{
\method{summary}{pretest}(object, ...)
}
\value{
No return value, called for summary.
}
\description{
Summary function for pretest
}
\keyword{internal}
