% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-SpotIV.R
\name{summary.SpotIV}
\alias{summary.SpotIV}
\title{Summary of SpotIV}
\usage{
\method{summary}{SpotIV}(object, ...)
}
\value{
No return value, called for summary.
}
\description{
Summary function for SpotIV and ProbitControl
}
\keyword{internal}
